requirements.txt
streamlit>=1.28.0
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
Pillow>=9.5.0
numpy>=1.24.0